bug -> session and bookstore failed login

+++ src/services/CreateBookstoreLoginService.ts
@@ -1,3 +1,4 @@
+import { compare } from 'bcrypt';
 import { sign } from 'jsonwebtoken';
 
 import { NewTokenDTO } from '../database/dtos/dtos';
@@ -8,12 +9,13 @@ import { AppError } from '../errors/AppError';
 import auth from '../settings/auth';
 
 type BookstoreDTO = {
+  adress?: string;
   name: string;
   isAdmin?: boolean;
 };
 
 class CreateBookstoreLoginService {
-  async execute({ name, isAdmin }: BookstoreDTO): Promise<NewTokenDTO> {
+  async execute({ name, adress, isAdmin }: BookstoreDTO): Promise<NewTokenDTO> {
     const { secret } = auth;
     const findByAdminRepository = new FindByAdminRepository();
     const adminExists = await findByAdminRepository.findByAdmin({ isAdmin });
@@ -27,15 +29,26 @@ class CreateBookstoreLoginService {
       throw new AppError('Bookstore not exists!', 404);
     }
 
+    /*   const adressMatch = await compare(adress, nameExists.adress);
+    console.log('adressMatch', adressMatch);
+    if (!adressMatch) {
+      throw new AppError('Incorrect');
+    } */
+
     const newToken = sign({ name }, secret, {
       subject: nameExists.id,
       expiresIn: '1d',
     });
+    console.log('newToken, ', newToken);
+    if (!newToken) {
+      throw new AppError('login failed, contact support', 401);
+    }
 
     const tokensRepository = new TokensRepository();
     const tokenConflict = await tokensRepository.findByBookstoreId({
       bookstoreId: nameExists.id,
     });
+    console.log('tokenConflict', tokenConflict);
     if (tokenConflict) {
       await tokensRepository.delete({ bookstoreId: nameExists.id });
     }
